19 feature request int add option to control output format

+++ src/RGen.Domain/Generating/IGenerator.cs
@@ -1,12 +1,13 @@
 ﻿namespace RGen.Domain.Generating;
 
+
 public interface IGenerator
 {
-//TODO: Refactor to use parameter-object instead of list of parameters
+//TODO: #42: Refactor to use parameter-object instead of list of parameters
 	IRandomValues Generate(
 		int numberOfElements,
 		int numberOfSets,
 		int? lengthOfElement,
-		int? min,
-		int? max);
+		ulong? min,
+		ulong? max);
 }
